functions/requirements.txt
anyio==4.0.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
cloudevents==1.9.0
coloredlogs==15.0.1
cryptography==41.0.3
deprecation==2.1.0
fastapi==0.99.1
filelock==3.12.3
firebase-admin==6.2.0
firebase-functions==0.1.1
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fsspec==2023.9.0
functions-framework==3.4.0
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.23.4
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.13.0
google-cloud-tasks==2.14.2
google-crc32c==1.5.0
google-events==0.10.0
google-generativeai==0.1.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.7
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
numpy==1.25.2
onnxruntime==1.15.1
overrides==7.4.0
packaging==23.1
pandas==2.1.0
posthog==3.0.2
proto-plus==1.22.3
protobuf==4.24.3
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
pyparsing==3.1.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
